feat: Add release_actions fastlane plugin
Adds a fastlane plugin to be hosted for usage on the iOS and Android
projects. Fastlane is a continuous delivery toolkit with pre-defined
actions to make it easier to build and deploy applications and
libraries. These actions and helper libraries handle various utility
functions such as parsing Conventional Commits, writing changelogs, and
bumping a version string as per the Semantic Versioning specification.

Co-authored-by: John Pignata <john@pignata.com>

Co-authored-by: Ivan Artemiev <29709626+iartemiev@users.noreply.github.com>