fix: permit all iot permissions in android iot test
The Android IoT tests arrange/create some resources. In order to delete
them, we need to include several more list/describe APIs. By the time
they are all added to the list, the IoT permissions list gets quite
long.

To ensure that the IoT stack is resilient to some changes in the client,
without needing to be continually updated/modified, we just move to
IoT:*.