fix(release_actions): Add factory method to Version
Currently iOS uses a prefix "v" for their semantic version strings. In
our calculations we need to account for this so that we can retrieve the
right diff from the repo. Previously we stripped this in our Git helper,
but that was insufficient as we also need the original tag. This adds a
factory method to Version that wrangles a tag into a proper Version
object.

Co-authored-by: Ivan Artemiev <29709626+iartemiev@users.noreply.github.com>